Factorize the CMake files and link against low-can
All target have to link to the same libraries so let's these lies into the main file.
Declare a variable holding the name of CAN low level project has been built and
installed. By default "rp-can-low-level". This is maily used to install the plugins
and configuration file to the correct place which has to be the place where the binding
was installed.

Change-Id: Id565bcdf8ab6063108a2f2b317cc2169808abb0e
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>